requirements.txt
streamlit
fastapi
uvicorn
PyPDF2
google-generativeai # Replace with latest version
python-dotenv
streamlit-extras # Replace with latest version
referencing
Pygments
attrs
protobuf
pillow
typing_extensions
packaging
python-multipart  